# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "plot api",
    "contact": {}
  },
  "host": "localhost:4000",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/behaviour": {
      "post": {
        "summary": "Create Behaviour",
        "tags": ["Behaviours"],
        "operationId": "CreateBehaviour",
        "deprecated": false,
        "produces": ["application/json; charset=utf-8"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CreateBehaviourRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CreateBehaviour"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "status": "Success",
                "message": "Create Skill LoCa Successfully "
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "64"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"40-4JBlX6SOq8hLDn8Z95Gtt5GyTVg\""
              },
              "Vary": {
                "type": "string",
                "default": "Accept-Encoding"
              },
              "Date": {
                "type": "string",
                "default": "Wed, 29 Sep 2021 12:28:04 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Behaviour",
        "tags": ["Behaviours"],
        "operationId": "DeleteBehaviour",
        "deprecated": false,
        "produces": ["application/json; charset=utf-8"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/DeleteBehaviourRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DeleteBehaviour"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Deleted Behaviour With Name Teaching"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "50"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"32-sq24iklXCnYJgXXIWr9DPiNimRg\""
              },
              "Vary": {
                "type": "string",
                "default": "Accept-Encoding"
              },
              "Date": {
                "type": "string",
                "default": "Wed, 29 Sep 2021 13:41:55 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      }
    },
...
    "/arm": {
      "post": {
        "summary": "create Arm",
        "tags": ["Arm"],
        "operationId": "createArm",
        "deprecated": false,
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createArmRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/arm/student/5f91165f12f9122febfd5dcf": {
      "get": {
        "summary": "get All Student in arm",
        "tags": ["Arm"],
        "operationId": "getAllStudentinarm",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/designation": {
      "post": {
        "summary": "create department",
        "tags": ["Desination"],
        "operationId": "createdepartment",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createdepartmentrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/department/edit": {
      "post": {
        "summary": "edit department",
        "tags": ["Department"],
        "operationId": "editdepartment",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/editdepartmentrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateBehaviourRequest": {
      "title": "CreateBehaviourRequest",
      "example": {
        "name": "LoCa",
        "school": "5f9183cad035174e94992476"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "school": {
          "type": "string"
        }
      },
      "required": ["name", "school"]
    },
    "CreateBehaviour": {
      "title": "CreateBehaviour",
      "example": {
        "status": "Success",
        "message": "Create Skill LoCa Successfully "
      },
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      },
      "required": ["status", "message"]
    },
...
    "DeleteBehaviourRequest": {
      "title": "DeleteBehaviourRequest",
      "example": {
        "behaviourId": "5fa26ec7c113d5001781c9f4",
        "school": "5f9183cad035174e94992476"
      },
      "type": "object",
      "properties": {
        "behaviourId": {
          "type": "string"
        },
        "school": {
          "type": "string"
        }
      },
      "required": ["behaviourId", "school"]
    },
    "DeleteBehaviour": {
      "title": "DeleteBehaviour",
      "example": {
        "message": "Deleted Behaviour With Name Teaching"
      },
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "required": ["message"]
    },
...
    "createArmRequest": {
      "title": "createArmRequest",
      "example": {
        "arm": "B",
        "classes": "5f902e9b9dafe11137c5ffb0",
        "school": "5f8c7ee1b9776e05f105a6db"
      },
      "type": "object",
      "properties": {
        "arm": {
          "type": "string"
        },
        "classes": {
          "type": "string"
        },
        "school": {
          "type": "string"
        }
      },
      "required": ["arm", "classes", "school"]
    },
...
    "createdepartmentrequest": {
      "title": "createdepartmentrequest",
      "example": {
        "name": "Teacher",
        "school": "5f8c7ee1b9776e05f105a6db"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "school": {
          "type": "string"
        }
      },
      "required": ["name", "school"]
    },
...
    "editdepartmentrequest": {
      "title": "editdepartmentrequest",
      "example": {
        "department": "5f9be928421b1a595d6747e5",
        "name": "javinokjjjjjjj",
        "school": "5f8c7ee1b9776e05f105a6db"
      },
      "type": "object",
      "properties": {
        "department": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "school": {
          "type": "string"
        }
      },
      "required": ["department", "name", "school"]
    },
...
  },
...
}